independent/requirements.txt
pdfminer.six
faiss-cpu
sentence-transformers
transformers
accelerate
bitsandbytes
jupyter
nbconvert
fastapi[all]
uvicorn[standard]
pyngrok
nest-asyncio
openai